refactor: #1294: Move match prediction debug models to debug client models namespace.
