+++ Atlas.Common.Public.Models/MatchPrediction/FrequencySetMetadata.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Atlas.MatchPrediction.ExternalInterface.Models.HaplotypeFrequencySet
+﻿namespace Atlas.Common.Public.Models.MatchPrediction
 {
     /// <summary>
     /// This is the data used to determine which frequency set to use, for both Donors and Patients.
