+++ Atlas.MatchPrediction.Test.Integration/IntegrationTests/MatchPrediction/IndividualSteps/HaplotypeFrequencySetSelection/HaplotypeFrequencySetSelectionTests.cs
@@ -6,6 +6,7 @@ using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 using System.Threading.Tasks;
+using Atlas.Common.Public.Models.MatchPrediction;
 using Atlas.HlaMetadataDictionary.Test.IntegrationTests.TestHelpers.FileBackedStorageStubs;
 
 namespace Atlas.MatchPrediction.Test.Integration.IntegrationTests.MatchPrediction.IndividualSteps.HaplotypeFrequencySetSelection
