+++ Atlas.MatchPrediction.Test/TestHelpers/Builders/MatchProbabilityInputs/DonorInputBuilder.cs
@@ -1,10 +1,9 @@
 using System.Collections.Generic;
 using System.Linq;
-using Atlas.Common.GeneticData.PhenotypeInfo;
 using Atlas.Common.Public.Models.GeneticData.PhenotypeInfo;
 using Atlas.Common.Public.Models.GeneticData.PhenotypeInfo.TransferModels;
+using Atlas.Common.Public.Models.MatchPrediction;
 using Atlas.Common.Test.SharedTestHelpers;
-using Atlas.MatchPrediction.ExternalInterface.Models.HaplotypeFrequencySet;
 using Builder = LochNessBuilder.Builder<Atlas.MatchPrediction.ExternalInterface.Models.MatchProbability.DonorInput>;
 
 namespace Atlas.MatchPrediction.Test.TestHelpers.Builders.MatchProbabilityInputs
