+++ Atlas.MatchPrediction/ExternalInterface/Models/HaplotypeFrequencySet/HaplotypeFrequencySetInput.cs
@@ -1,4 +1,6 @@
-﻿namespace Atlas.MatchPrediction.ExternalInterface.Models.HaplotypeFrequencySet
+﻿using Atlas.Common.Public.Models.MatchPrediction;
+
+namespace Atlas.MatchPrediction.ExternalInterface.Models.HaplotypeFrequencySet
 {
     public class HaplotypeFrequencySetInput
     {
