+++ Atlas.MatchPrediction/ExternalInterface/Models/MatchPredictionSteps/GenotypeLikelihood/GenotypeLikelihoodInput.cs
@@ -1,9 +1,7 @@
 ﻿using System.Collections.Generic;
-using Atlas.Common.GeneticData;
-using Atlas.Common.GeneticData.PhenotypeInfo;
 using Atlas.Common.Public.Models.GeneticData;
 using Atlas.Common.Public.Models.GeneticData.PhenotypeInfo;
-using Atlas.MatchPrediction.ExternalInterface.Models.HaplotypeFrequencySet;
+using Atlas.Common.Public.Models.MatchPrediction;
 
 namespace Atlas.MatchPrediction.ExternalInterface.Models.MatchPredictionSteps.GenotypeLikelihood
 {
