+++ Atlas.MatchPrediction/Services/MatchProbability/MatchProbabilityService.cs
@@ -13,6 +13,7 @@ using FluentValidation;
 using System.Linq;
 using System.Threading.Tasks;
 using Atlas.Common.Public.Models.GeneticData;
+using Atlas.Common.Public.Models.MatchPrediction;
 
 // ReSharper disable ParameterTypeCanBeEnumerable.Local
 // ReSharper disable SuggestBaseTypeForParameter
